requirements.txt
requests==2.31.0
gspread==5.12.0
google-auth==2.38.0  # Заменяем oauth2client
pybit==5.5.0
python-telegram-bot==20.6


